+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - allow "key" to be passed as a parameter
 - allow "room_id" to be passed as a parameter or with the data
 - rename "text" to "body".
+- Publish releases also on github from github actions
 
 ## [3.1.1] - 2021-07-18