+++ CHANGELOG.md
@@ -47,7 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2020-02-14
 - First release with matrix-client & http.server
 
-[Unreleased]: https://github.com/nim65s/matrix-webhook/compare/v3.1.1...devel
+[Unreleased]: https://github.com/nim65s/matrix-webhook/compare/v3.1.1...master
 [3.1.1]: https://github.com/nim65s/matrix-webhook/compare/v3.1.0...v3.1.1
 [3.1.0]: https://github.com/nim65s/matrix-webhook/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/nim65s/matrix-webhook/compare/v2.0.0...v3.0.0